test(slack): settle identity delivery proof
