+++ extensions/slack/src/monitor.test-helpers.ts
@@ -25,8 +25,6 @@ type SlackProviderMonitor = (params: {
   setStatus?: (next: Record<string, unknown>) => void;
 }) => Promise<unknown>;
 type SlackStartupAuthClientFactory = typeof import("./client.js").createSlackStartupAuthClient;
-type SlackChannelInboundDispatch =
-  typeof import("openclaw/plugin-sdk/channel-inbound").dispatchChannelInboundTurn;
 
 const SLACK_INGRESS_LIFECYCLE_CONTEXT_KEY = "openclawIngressLifecycle";
 
@@ -79,7 +77,6 @@ type SlackTestState = {
   >;
   socketModeLogger?: { error: (...args: unknown[]) => void };
   createSlackStartupAuthClientMock: Mock<SlackStartupAuthClientFactory>;
-  channelInboundDispatchOnce?: SlackChannelInboundDispatch;
 };
 
 // globalThis-backed singleton: with isolate=false, a vi.resetModules() in any
@@ -106,7 +103,6 @@ const slackTestState: SlackTestState = vi.hoisted(() => {
     resolveSlackUserAllowlistMock: vi.fn(),
     socketModeLogger: undefined,
     createSlackStartupAuthClientMock: vi.fn(),
-    channelInboundDispatchOnce: undefined,
   } as SlackTestState;
   return globalState["__slackTestState"];
 });
@@ -117,8 +113,11 @@ export function useSlackStartupAuthClientOnce(factory: SlackStartupAuthClientFac
   slackTestState.createSlackStartupAuthClientMock.mockImplementationOnce(factory);
 }
 
-export function useSlackChannelInboundDispatchOnce(dispatch: SlackChannelInboundDispatch): void {
-  slackTestState.channelInboundDispatchOnce = dispatch;
+export async function runSlackHandlerWithDispatch(
+  handler: SlackHandler,
+  args: unknown,
+): Promise<void> {
+  await handler(withSlackDispatchLifecycle(args));
 }
 
 type SlackClient = {
@@ -279,9 +278,12 @@ async function runSlackEventOnce(
   const handler = await getSlackHandlerOrThrow(name);
   // Normal Bolt handlers return after queue admission. Terminal-state tests use the
   // durable-ingress lifecycle so this helper can await the actual dispatch boundary.
-  const handlerArgs = opts?.awaitDispatch ? withSlackDispatchLifecycle(args) : args;
   try {
-    await handler(handlerArgs);
+    if (opts?.awaitDispatch) {
+      await runSlackHandlerWithDispatch(handler, args);
+    } else {
+      await handler(args);
+    }
   } finally {
     await stopSlackMonitor({ controller, run });
   }
@@ -348,7 +350,6 @@ export function resetSlackTestState(config: Record<string, unknown> = defaultSla
   slackTestState.config = config;
   slackTestState.appConstructorArgs = undefined;
   slackTestState.socketModeLogger = undefined;
-  slackTestState.channelInboundDispatchOnce = undefined;
   slackTestState.appStartMock.mockReset().mockResolvedValue(undefined);
   slackTestState.appStopMock.mockReset().mockResolvedValue(undefined);
   slackTestState.interactionRegistrations.length = 0;
@@ -413,15 +414,8 @@ vi.mock("openclaw/plugin-sdk/channel-inbound", async (importOriginal) => {
     slackTestState.replyMock(...args) as ReturnType<ReplyResolver>;
   return {
     ...actual,
-    dispatchChannelInboundTurn: (params: DispatchParams) => {
-      const enrichedParams = { ...params, replyResolver };
-      const dispatchOnce = slackTestState.channelInboundDispatchOnce;
-      if (dispatchOnce) {
-        slackTestState.channelInboundDispatchOnce = undefined;
-        return dispatchOnce(enrichedParams);
-      }
-      return actual.dispatchChannelInboundTurn(enrichedParams);
-    },
+    dispatchChannelInboundTurn: (params: DispatchParams) =>
+      actual.dispatchChannelInboundTurn({ ...params, replyResolver }),
   };
 });
